Fixed sorting and stats display
Fixed an issue with the stats display where the app would crash if
there was above a certain number of episodes, due to a maximum data
size that can be sent in an intent. Reduced the size of each element
to increase the number of elements that can be sent. This will still
be a problem in the future.