Fix crash caused by new medias with no DB entries
A crash was seen where a new media was added and incremented,
and then another new media was added. The root cause is that
the first new media increment has not updated the DB, but the
local count is still incremented. Adding the second new media
causes the comparison logic to run, leading to an index out of
bounds exception on the first media when trying to access the
latest episode.