Refactor import and exporting data
Refactored import and exporting data to use the file picker interface to
select files to read and write. This allows for the removal of the
storage shenanigans and permissions, and allows for selecting whichever
file and location to read or write to. Also fixed a bug in the
IonDataManager due to not correctly importing MediaDatas that were
marked as COMPLETE.